==> notebooks/01_load_and_validate_raw_data.ipynb <==
from pathlib import Path
import requests
import pandas as pd

def download_one_file_of_raw_data(year: int, month: int) -> Path:
    """"""
    URL = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet"
    response = requests.get(URL)
    
    if response.status_code == 200:
        # extract the contents of the response
        path = f"../data/raw/rides_{year}-{month:02d}.parquet"
        # save the contents to a file
        open(path, 'wb').write(response.content)
        return path
    else:
        raise Exception(f"{URL} is not available.")

download_one_file_of_raw_data(year = 2022, month = 1)

Load the parquet file into pandas dataframe

rides_df = pd.read_parquet("../data/raw/rides_2022-01.parquet")

rides_df.head(20)

rides_df_copy = rides_df.copy()

rides_df = rides_df[["tpep_pickup_datetime", "PULocationID"]]


# rename column names
rides_df.rename(columns = {"tpep_pickup_datetime": "pickup_datetime", "PULocationID": "pickup_location_id"}, inplace = True)

rides_df

rides_df["pickup_datetime"].describe()

**Observation:** We notice that Jan month 2022 dataset has dates in the year 2008 and May month of 2022. We have to remove them
- first value = 2008-12-31 22:23:09
- last value = 2022-05-18 20:41:57


rides_df = rides_df[rides_df["pickup_datetime"] >= "2022-01-01"]
rides_df = rides_df[rides_df["pickup_datetime"] <= "2022-02-01"]

rides_df["pickup_datetime"].describe()

**Observations:**

- Now we only have data of Jan 2022

rides_df.to_parquet("../data/transformed/validated_rides_2022_01.parquet")
