==> notebooks/02_transform_raw_data_into_time_series_data.ipynb <==
import pandas as pd
from tqdm import tqdm
from typing import Optional, List
import plotly.express as px

rides_df = pd.read_parquet("../data/transformed/validated_rides_2022_01.parquet")

rides_df.head(10)

Since we are working on time series data, we create a new hour column that contains the hour of the day for each ride.

rides_df["pickup_hour"] = rides_df["pickup_datetime"].dt.floor("H")

rides_df

Count the number of rides per location ID per pickup hour

# size() counts the no. of rows per group
agg_rides = rides_df.groupby(["pickup_hour", "pickup_location_id"]).size().reset_index(name="rides_count")

agg_rides

agg_rides[agg_rides["rides_count"] == 0]

def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    """add zeros for missing pickup_hour and pickup_location_id combinations
    """
    
    location_ids = agg_rides["pickup_location_id"].unique()
    # full range of dates
    full_range = pd.date_range(start=agg_rides["pickup_hour"].min(), end=agg_rides["pickup_hour"].max(), freq="H")
    output = pd.DataFrame()
    for location_id in tqdm(location_ids):
        # keep only rides for this location_id
        agg_rides_i = agg_rides.loc[agg_rides["pickup_location_id"] == location_id, ["pickup_hour", "rides_count"]]
        # qucik way to add missing dates with 0 in a series
        agg_rides_i.set_index("pickup_hour", inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)
        
        # add back location_id column
        agg_rides_i["pickup_location_id"] = location_id
        
        output = pd.concat([output, agg_rides_i])
        
    # move the purchase_day from the index to dataframe column
    output = output.reset_index().rename(columns={"index": "pickup_hour"})
    
    return output

agg_rides_all_slots = add_missing_slots(agg_rides)

agg_rides_all_slots

def plot_rides(rides: pd.DataFrame, location_lst: Optional[List[int]] = None):
    """
    Plot time-series data"""
    rides_to_plot = rides[rides["pickup_location_id"].isin(location_lst)] if location_lst else rides_df
    
    fig = px.line(rides_to_plot, x="pickup_hour", y="rides_count", color="pickup_location_id")
    fig.show()

plot_rides(agg_rides_all_slots, location_lst=[43])

agg_rides_all_slots.to_parquet("../data/transformed/ts_data_2022_01.parquet")
