==> notebooks/09_lightgbm_model_with_feature_engineering.ipynb <==
import os

%pwd

os.chdir('../')

%pwd

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")

df

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date = datetime(2022, 6, 1, 0, 0, 0),
    target_column_name = "target_rides_next_hour"
)


print(f"X_train shape: {X_train.shape}")
print(f"y_train shape: {y_train.shape}")
print(f"X_test shape: {X_test.shape}")
print(f"y_test shape: {y_test.shape}")

### Feature Engineering

- Perform data transformations to the data

### Method 1: Data Transformation using scikit learn's FunctionTransformer API

- add new column in the X_train


def average_rides_last_4_weeks(X: pd.DataFrame) -> pd.DataFrame:
    """Adds 1 column with the avergae rides from
    - 7 days ago
    - 14 days ago
    - 21 days ago
    - 28 days ago
    
    Args:
        X (pd.DataFrame): _description_
    
    Returns:
        pd.DataFrame: _description_
    """
    
    X["average_rides_last_4_weeks"] = 0.25 * (
        X[f"rides_previous_{7 * 24}_hour"]  + \
        X[f"rides_previous_{2 * 7 * 24}_hour"] + \
        X[f"rides_previous_{3 * 7 * 24}_hour"] + \
        X[f"rides_previous_{4 * 7 * 24}_hour"]
    )
    
    return X

Use scikit learn **FunctionTransformer** to add the average_rides_last_4_weeks() method logic in the feature engineering pipeline

from sklearn.preprocessing import FunctionTransformer

add_feature_average_rides_last_4_weeks = FunctionTransformer(
    average_rides_last_4_weeks, validate=False
    )

Fit_Transform() --> The below step will add a new column called "average_rides_last_4_weeks" to the X_train

add_feature_average_rides_last_4_weeks.fit_transform(X_train)

### Method 2: Data tranformation using custom class (most customizable way ot define data transformation)

- Extract hour and day of week from "pickup_hour" column

from sklearn.base import BaseEstimator, TransformerMixin

class TemporalFeatureEngineering(BaseEstimator, TransformerMixin):
    """
    """
    
    def fit(self, X, y = None):
        return self
    
    
    def transform(self, X, y = None):
        
        X_ = X.copy()
        
        # Generate numeric columns from datetime
        X_["hour"] = X_["pickup_hour"].dt.hour
        X_["day_of_week"] = X_["pickup_hour"].dt.dayofweek
        
        # remove pickup_hour column since it's a datetime column and lightgbm only accepts numeric columns
        return X_.drop(columns=["pickup_hour"])
        

add_temporal_features = TemporalFeatureEngineering()
add_temporal_features.fit_transform(X_train)

### After feature Engineering, put all the components in the scikit-learn pipeline

import lightgbm as lgb
from sklearn.pipeline import make_pipeline

pipeline = make_pipeline(
    add_feature_average_rides_last_4_weeks,
    add_temporal_features,
    lgb.LGBMRegressor()
)

pipeline.fit(X_train, y_train)

Predictions

predictions = pipeline.predict(X_test)

Evaluate

from sklearn.metrics import mean_absolute_error

test_mar = mean_absolute_error(y_test, predictions)
print(f"Test MAE of lightGBM with feature engineering: {test_mar:.4f}")

- Feature engineering further reduced our MAE score from 2.62 to 2.59 (in 8th notebook).  So, this is an improvement to the lightBGM model

- We can further improve the model by hyperparameter tuning

### References:

- https://towardsdatascience.com/how-to-quickly-design-advanced-sklearn-pipelines-3cc97b59ce16
