==> notebooks/10_lightgbm_model_with_hyperparameter_tuning.ipynb <==
import os

%pwd

os.chdir('../')

%pwd

import warnings
warnings.filterwarnings('ignore')

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")
df

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date = datetime(2022, 6, 1, 0, 0, 0),
    target_column_name = "target_rides_next_hour"
)


print(f"X_train shape: {X_train.shape}")
print(f"y_train shape: {y_train.shape}")
print(f"X_test shape: {X_test.shape}")
print(f"y_test shape: {y_test.shape}")

### Define function to map set of hyperparameters to a final validation error

import numpy as np
from sklearn.model_selection import KFold, TimeSeriesSplit
from sklearn.pipeline import make_pipeline
from sklearn.metrics import mean_absolute_error
from src.model import get_pipeline

# for hyperparameter tuning
import optuna 


def objective(trial: optuna.trial.Trial) -> float:
    """ 
    Given a set of hyperparameters, it trains a model and computes an average 
    validation error based on a TimeSeriesSplit cross-validation strategy.
    
    """
    
    # pick hyperparameters
    hyperparams = {
        "metric": "mae",
        "verbose": -1,
        "num_leaves": trial.suggest_int("num_leaves", 2, 256),
        "feature_fraction": trial.suggest_float("feature_fraction", 0.2, 1.0),
        "bagging_fraction": trial.suggest_float("bagging_fraction", 0.2, 1.0),
        "min_child_samples": trial.suggest_int("min_child_samples", 3, 100)
    }
    
    # split the dataset into n splits
    tss = TimeSeriesSplit(n_splits = 5) # this is train (n-1) models. So, 5 - 1 = 4 models
    scores = []
    
    for train_idx, val_idx in tss.split(X_train):
        
        # split train and validation sets for hyperparameter tuning
        X_train_, X_val_ = X_train.iloc[train_idx, :], X_train.iloc[val_idx, :]
        y_train_, y_val_ = y_train[train_idx], y_train[val_idx]
        
        # train the model
        pipeline = get_pipeline(**hyperparams)
        pipeline.fit(X_train_, y_train_)
        
        # evaluate the model
        y_pred = pipeline.predict(X_val_)
        mae = mean_absolute_error(y_val_, y_pred)
        
        scores.append(mae)
    
    # return the average validation error
    return np.array(scores).mean()

Now, minimize the above objective function using Optuna study API

study = optuna.create_study(direction = "minimize")
study.optimize(objective, n_trials = 5)

Extract the best Hyperparameters that Optuna found

best_params = study.best_trial.params
print(f"Best params using Optuna: {best_params}")

### Retrain the model using the best hyperparameters

pipeline = get_pipeline(**best_params)
pipeline.fit(X_train, y_train)

predictions = pipeline.predict(X_test)
test_mae = mean_absolute_error(y_test, predictions)
print(f"Test MAE of lightGBM after hyperparameter tuning: {test_mae:.4f}")

### Plot

from src.plot import plot_one_sample

plot_one_sample(
    features = X_test,
    target = y_test,
    example_id = 2979,
    predictions = pd.Series(predictions)
)


plot_one_sample(
    features = X_test,
    target = y_test,
    example_id = 3979,
    predictions = pd.Series(predictions)
)

### The LigthGBM model predictions are pretty good
