==> notebooks/11_backfill_feature_store.ipynb <==
import os

%pwd

os.chdir('../')

%pwd

HOPSWORKS_PROJECT_NAME = 'taxi_demand_pred'

### Method to load .env file containing the HOPSWORKS API key

from src.paths import PARENT_DIR
from dotenv import load_dotenv

# load key-value pairs from .env file located in the parent directory
load_dotenv(PARENT_DIR / ".env") 

HOPSWORKS_API_KEY = os.environ["HOPSWORKS_API_KEY"]

### Fetch Historical data for 2022 - 2023

from datetime import datetime
import pandas as pd
from src.data_processing import load_raw_data

from_year = 2022
to_year = 2023
print(f"Downloading raw data from {from_year} to {to_year}")

rides = pd.DataFrame()

for year in range(from_year, to_year + 1):
        
    # Download data for the whole year
    rides_one_year = load_raw_data(year)
    
    # appennd rows
    rides = pd.concat([rides, rides_one_year])

Check total rows in the data

print(f"Raw data size of both 20222 and 220223 year: {len(rides)}")

### Transform the raw data into time series data

from src.data_processing import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)

### Save the data in feature store

- Following are the steps

import hopsworks

#### Connect to project on hopsworks

project = hopsworks.login(
    project = HOPSWORKS_PROJECT_NAME,
    api_key_value = HOPSWORKS_API_KEY
)

#### Connect to the feature store

feature_store = project.get_feature_store()

#### Now, save the data into the feature store

- to save data, we need to use feature groups which is the API for writing data into feature store

- To use feature group, we need to set it's name and version

FEATURE_GROUP_NAME = "time_series_hourly_feature_group"
FEATURE_GROUP_VERSION = 1

#### Now, create the feature group

feature_group = feature_store.get_or_create_feature_group(
    name = FEATURE_GROUP_NAME,
    version = FEATURE_GROUP_VERSION,
    description = "Time series data at hourly frequency",
    primary_key = ["pickup_location_id", "pickup_hour"],
    event_time = "pickup_hour"
)

#### Finally save the data into the feature store

feature_group.insert(ts_data, write_options = {"wait_for_job": False})
