==> notebooks/12_feature_pipeline.ipynb <==
import os

%pwd

os.chdir("../")

%pwd

### This note book will implement the feature pipeline.

- This is run automatically every hour and is going to fetch a batch of recent data, transform it and save it to the feature store

import src.config as config
import pandas as pd
from datetime import datetime, timedelta

current_date = pd.to_datetime(datetime.utcnow()).floor("H")
print(f"current_date: {current_date}")


# we fetch the raw data for the last 28 days, to add redundancy to our data pipeline.
# redundancy coz in case pipeline failes for a certain hour, it's not gonna break things completely
fetch_data_to = current_date
fetch_data_from = current_date - timedelta(days = 28)

print(f"fetch_data_from: {fetch_data_from}")
print(f"fetch_data_to: {fetch_data_to}")

### Now, we need to fetch recent data

- We don't have access to NYC assosiation data warehouse, so we will simulate a call to data warehouse

from src.data_processing import load_raw_data

def fetch_batch_raw_data(from_date: datetime, to_date: datetime) -> pd.DataFrame:
    """Simulate production data by sampling historical data from 52 weeks ago(i.e. 1 year ago)

    Args:
        from_date (datetime): _description_
        to_date (datetime): _description_

    Returns:
        pd.DataFrame: dataframe of historical data of that time period
    """
    
    from_date_ = from_date - timedelta(days = 7 * 52)
    to_date_ = to_date - timedelta(days = 7 * 52)
    print(f'{from_date=}, {to_date_=}')
    
    # download 2 files from website
    rides = load_raw_data(year = from_date_.year, month_lst = from_date_.month)
    rides = rides[rides.pickup_datetime >= from_date_]
    rides_2 = load_raw_data(year = to_date_.year, month_lst = to_date_.month)
    rides_2 = rides_2[rides_2.pickup_datetime < to_date_]
    
    
    rides = pd.concat([rides, rides_2])
    
    # shift the data by 1 year to pretend this is recent data
    rides["pickup_datetime"] += timedelta(weeks = 7 * 52)
    
    rides.sort_values(by = ["pickup_location_id" ,"pickup_datetime"], inplace = True)
    
    
    return rides

### Download the raw data of 1 yr back

rides = fetch_batch_raw_data(from_date = fetch_data_from, to_date = fetch_data_to)

### Transform the raw data to time series data

from src.data_processing import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)

### Connect to feature store and feature group where we want to store the data

import hopsworks

# connect to the project
project = hopsworks.login(
    project = config.HOPSWORKS_PROJECT_NAME,
    api_key_value = config.HOPSWORKS_API_KEY
)

# connect to the feature store
feature_store = project.get_feature_store()

# connect to the feature group
feature_group = feature_store.get_or_create_feature_group(
    name = config.FEATURE_GROUP_NAME,
    version = config.FEATURE_GROUP_VERSION,
    description = "Time-series data hourly frequency",
    primary_key = ["pickup_location_id", "pickup_hour"],
    event_time = "pickup_hour"
)

### Now, insert the data into the feature group

feature_group.insert(ts_data, write_options = {"wait_for_job": False})

### In the next notebook, we will automate this execution by running this notebook  automatically every hour. We use github actions for this task
