==> notebooks/13_model_training_pipeline.ipynb <==
import os

%pwd

os.chdir("../")

%pwd

### Read the data from Feature Store

- Save the model in Hopsworks model registry which can be used to generate predictions

import src.config as config
import hopsworks

# connect to the project
project = hopsworks.login(
    project = config.HOPSWORKS_PROJECT_NAME,
    api_key_value = config.HOPSWORKS_API_KEY
)

# connect to feature store
feature_store = project.get_feature_store()

# connect to feature group
feature_group = feature_store.get_feature_group(
    name = config.FEATURE_GROUP_NAME,
    version = config.FEATURE_GROUP_VERSION
)

### To read data from feature store, we need to create feature view

- With feature view, you define how to fetch data from potentially many different feature groups and merge it to get the final data set

- Feature view is metadata. It's just a definition for hopsworks to understand how we want to fetch the data

# create a feature view (if it doesn't exist yet)
# This feature view only uses on feature groups, so the query is trivial

try:
    # create feature view if it doesn't exist yet
    feature_store.create_feature_view(
        name = config.FEATURE_VIEW_NAME,
        version = config.FEATURE_VIEW_VERSION,
        query = feature_group.select_all()
    )
except:
    print(f"Feature view already exists. Skipping Creation")
    

# get feature view
feature_view = feature_store.get_feature_view(
    name = config.FEATURE_VIEW_NAME,
    version = config.FEATURE_VIEW_VERSION
)

### Now, if you want to fetch the data, we need to create training data

ts_data = feature_view.training_data(
    description = "Time series hourly taxi rides",
)
